app/requirements.txt
Django>=3.2,<4.0
factory_boy>=3.2.0
Faker>=13.3.4
